django-ossature-ld/pyproject.toml
[project]
name = "django-ossature-ld"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.6",
    "django-allauth-ui>=1.8.1",
    "django-allauth[mfa]>=65.11.2",
    "django-countries[pyuca]>=7.6.1",
    "django-debug-toolbar>=6.0.0",
    "django-encrypted-model-fields>=0.6.5",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-import-export[all]>=4.3.9",
    "django-model-utils>=5.0.0",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-recaptcha>=4.1.0",
    "django-robots>=6.1",
    "django-rosetta>=0.10.2",
    "pwned-passwords-django>=5.2.0",
]
